django_app/requirements.txt
keras
Django
Pillow
numpy
tensorflow
pandas
sklearn
matplotlib
